+++ tests/testthat/test-correction.R
@@ -19,5 +19,5 @@ attributes(coefs_bc) <- NULL
 mod_bc <- glm(conc ~ lot*log(u), data = clotting, family = Gamma, method = "brglmFit", type = "correction")
 
 test_that("bias corrected estimates computed using enrichwith are the same as those when having type = 'correction' in brglmFit", {
-    expect_equal(unname(coefs_bc), unname(coef(mod_bc, model = "full")), check.attributes = FALSE, tolerance = 1e-10)
+    expect_equal(unname(coefs_bc), unname(coef(mod_bc, model = "full")), check.attributes = FALSE, tolerance = 1e-06)
 })